+++ controllers/call/controller.go
@@ -30,11 +30,9 @@ import (
 	"github.com/carv-ics-forth/frisbee/pkg/lifecycle"
 	"github.com/carv-ics-forth/frisbee/pkg/scheduler"
 	"github.com/go-logr/logr"
-	cmap "github.com/orcaman/concurrent-map"
 	"github.com/pkg/errors"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/types"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -58,14 +56,10 @@ type Controller struct {
 	ctrl.Manager
 	logr.Logger
 
-	gvk schema.GroupVersionKind
-
 	view *lifecycle.Classifier
 
 	// executor is used to run commands directly into containers
 	executor kubexec.Executor
-
-	regionAnnotations cmap.ConcurrentMap
 }
 
 // Reconcile is part of the main kubernetes reconciliation loop which aims to
@@ -146,37 +140,33 @@ func (r *Controller) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Resu
 	case v1alpha1.PhasePending:
 		nextJob := call.Status.ScheduledJobs + 1
 
-		/*
-			If all jobs are scheduled but are not in the Running phase, they may be in the Pending phase.
-			In both cases, we have nothing else to do but waiting for the next reconciliation cycle.
-		*/
+		//	If all jobs are scheduled but are not in the Running phase, they may be in the Pending phase.
+		//	In both cases, we have nothing else to do but waiting for the next reconciliation cycle.
 		if call.Spec.SuspendWhen == nil && (nextJob >= len(call.Status.QueuedJobs)) {
 			return common.Stop(r, req)
 		}
 
 		// Get the next scheduled job
-		{
-			hasJob, nextTick, err := scheduler.Schedule(log, &call, scheduler.Parameters{
-				State:            *r.view,
-				ScheduleSpec:     call.Spec.Schedule,
-				LastScheduleTime: call.Status.LastScheduleTime,
-				ExpectedTimeline: call.Status.ExpectedTimeline,
-				JobName:          call.GetName(),
-				ScheduledJobs:    call.Status.ScheduledJobs,
-			})
-			if err != nil {
-				return lifecycle.Failed(ctx, r, &call, errors.Wrapf(err, "scheduling error"))
+		hasJob, nextTick, err := scheduler.Schedule(log, &call, scheduler.Parameters{
+			State:            *r.view,
+			ScheduleSpec:     call.Spec.Schedule,
+			LastScheduleTime: call.Status.LastScheduleTime,
+			ExpectedTimeline: call.Status.ExpectedTimeline,
+			JobName:          call.GetName(),
+			ScheduledJobs:    call.Status.ScheduledJobs,
+		})
+		if err != nil {
+			return lifecycle.Failed(ctx, r, &call, errors.Wrapf(err, "scheduling error"))
+		}
+
+		if !hasJob {
+			r.Logger.Info("Nothing to do right now. Requeue request.", "nextTick", nextTick)
+
+			if nextTick.IsZero() {
+				return common.Stop(r, req)
 			}
 
-			if !hasJob {
-				r.Logger.Info("Nothing to do right now. Requeue request.", "nextTick", nextTick)
-
-				if nextTick.IsZero() {
-					return common.Stop(r, req)
-				}
-
-				return common.RequeueAfter(r, req, time.Until(nextTick))
-			}
+			return common.RequeueAfter(r, req, time.Until(nextTick))
 		}
 
 		// Build the job in kubernetes
@@ -350,23 +340,18 @@ func (r *Controller) Finalize(obj client.Object) error {
 */
 
 func NewController(mgr ctrl.Manager, logger logr.Logger) error {
-	r := &Controller{
-		Manager:           mgr,
-		Logger:            logger.WithName("call"),
-		gvk:               v1alpha1.GroupVersion.WithKind("Call"),
-		view:              &lifecycle.Classifier{},
-		executor:          kubexec.NewExecutor(mgr.GetConfig()),
-		regionAnnotations: cmap.New(),
+	reconciler := &Controller{
+		Manager:  mgr,
+		Logger:   logger.WithName("call"),
+		view:     &lifecycle.Classifier{},
+		executor: kubexec.NewExecutor(mgr.GetConfig()),
 	}
 
-	var (
-		call    v1alpha1.Call
-		vobject v1alpha1.VirtualObject
-	)
+	gvk := v1alpha1.GroupVersion.WithKind("Call")
 
 	return ctrl.NewControllerManagedBy(mgr).
-		For(&call).
+		For(&v1alpha1.Call{}).
 		Named("call").
-		Owns(&vobject, watchers.NewWatchWithRangeAnnotations(r, r.gvk)).
-		Complete(r)
+		Owns(&v1alpha1.VirtualObject{}, watchers.Watch(reconciler, gvk)).
+		Complete(reconciler)
 }